# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Challenge 05: Products in shop",
    "description": "This is the documentation of the API created for Challenge Chapter 5 on Binar Academy. If you find an error or have something you want to ask, then contact me on whatsapp.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "brahmastabagus@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1"
    }
  ],
...
  "paths": {
...
    "/users/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user",
        "description": "Get user by id.",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success get users by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int64",
                          "example": 10
                        },
                        "username": {
                          "type": "string",
                          "example": "admin"
                        },
                        "password": {
                          "type": "string",
                          "example": "admin123"
                        },
                        "role": {
                          "type": "string",
                          "default": "admin",
                          "enum": [
                            "superadmin",
                            "admin",
                            "customer"
                          ]
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2023-05-04T14:20:27.537Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2023-05-04T14:20:27.537Z"
                        },
                        "shop": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string",
                              "example": "Bagus shop"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Data not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "400": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res400"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update user",
        "description": "This is done to be able to login.",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Updated user object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Update user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 2 telah berhasil terupdate"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Username is already",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Username admin sudah ada"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res400"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete user",
        "description": "Delete user by id.",
        "operationId": "deleteUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success delete user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 3 telah berhasil terhapus"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Data not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "400": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res400"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      }
    },
...
    "/products/{id}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Get product",
        "description": "Get product by id.",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success get product by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Products"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Data not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "400": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res400"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "product"
        ],
        "summary": "Update product",
        "description": "This is done to be able to login.",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Created product object",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Update product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 2 telah berhasil terupdate"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Price must be above 10.000",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "Harga harus diatas 10.000"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Data not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "400": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res400"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "product"
        ],
        "summary": "Delete product",
        "description": "Delete product by id.",
        "operationId": "deleteProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success delete product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Data dengan index 3 telah berhasil terhapus"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Data not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "400": {
            "description": "Failed bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Res400"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "example": "admin123"
          },
          "role": {
            "type": "string",
            "default": "admin",
            "enum": [
              "superadmin",
              "admin",
              "user"
            ]
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Kaos kaki"
          },
          "price": {
            "type": "integer",
            "example": 5000
          },
          "stock": {
            "type": "integer",
            "example": 40
          },
          "imageUrl": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "name",
          "price",
          "stock",
          "imageUrl"
        ]
      },
      "Products": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "name": {
            "type": "string",
            "example": "Kaos kaki"
          },
          "price": {
            "type": "integer",
            "example": 5000
          },
          "stock": {
            "type": "integer",
            "example": 40
          },
          "imageUrl": {
            "type": "string",
            "example": "https://ik.imagekit.io/brahmastabagus/IMG-1683214309867_g9PFH5Dme.jpeg"
          },
          "createdAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-05-04T14:20:27.537Z"
          },
          "shopId": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "NotFound": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failed"
          },
          "message": {
            "type": "string",
            "example": "Data dengan id 100, tidak ditemukan"
          }
        }
      },
      "Res400": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failed"
          },
          "message": {
            "type": "string",
            "example": "Cannot proccess"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}